+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/config/ApplicationStartupConfig.java
@@ -19,21 +19,29 @@
 
 package org.eclipse.tractusx.traceability.common.config;
 
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.eclipse.tractusx.traceability.policies.domain.PolicyRepository;
+import org.eclipse.tractusx.traceability.assets.application.base.service.AssetBaseService;
+import org.eclipse.tractusx.traceability.assets.domain.base.model.AssetBase;
+import org.eclipse.tractusx.traceability.contracts.application.service.ContractService;
+import org.eclipse.tractusx.traceability.contracts.domain.model.ContractAgreement;
+import org.eclipse.tractusx.traceability.contracts.domain.model.ContractType;
 import org.eclipse.tractusx.traceability.notification.application.contract.model.CreateNotificationContractRequest;
 import org.eclipse.tractusx.traceability.notification.application.contract.model.NotificationMethod;
 import org.eclipse.tractusx.traceability.notification.application.contract.model.NotificationType;
 import org.eclipse.tractusx.traceability.notification.domain.contract.EdcNotificationContractService;
+import org.eclipse.tractusx.traceability.policies.domain.PolicyRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
+import java.time.Instant;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.stream.Stream;
 
 import static org.eclipse.tractusx.traceability.common.config.ApplicationProfiles.NOT_INTEGRATION_TESTS;
 
@@ -41,9 +49,25 @@ import static org.eclipse.tractusx.traceability.common.config.ApplicationProfile
 @Slf4j
 @Component
 @Profile(NOT_INTEGRATION_TESTS)
-@RequiredArgsConstructor
 public class ApplicationStartupConfig {
     private final PolicyRepository policyRepository;
+    private final AssetBaseService asPlannedService;
+    private final AssetBaseService asBuiltService;
+    private final ContractService contractService;
+
+    @Autowired
+    public ApplicationStartupConfig(PolicyRepository policyRepository,
+                                    @Qualifier("assetAsBuiltServiceImpl") AssetBaseService asPlannedService,
+                                    @Qualifier("assetAsPlannedServiceImpl") AssetBaseService asBuiltService,
+                                    EdcNotificationContractService edcNotificationContractService,
+                                    ContractService contractService) {
+        this.policyRepository = policyRepository;
+        this.asPlannedService = asPlannedService;
+        this.asBuiltService = asBuiltService;
+        this.edcNotificationContractService = edcNotificationContractService;
+        this.contractService = contractService;
+    }
+
     private final EdcNotificationContractService edcNotificationContractService;
     private static final List<CreateNotificationContractRequest> NOTIFICATION_CONTRACTS = List.of(
             new CreateNotificationContractRequest(NotificationType.QUALITY_ALERT, NotificationMethod.UPDATE),
@@ -81,5 +105,45 @@ public class ApplicationStartupConfig {
         });
         executor.shutdown();
     }
+
+    @EventListener(ApplicationReadyEvent.class)
+    public void insertIntoContractAgreements() {
+        ExecutorService executor = Executors.newSingleThreadExecutor();
+        executor.execute(() -> {
+            log.info("on ApplicationReadyEvent insert into contracts.");
+            try {
+                List<AssetBase> asBuilt = asBuiltService.findAll();
+                List<AssetBase> asPlanned = asPlannedService.findAll();
+
+                List<ContractAgreement> contractAgreementIdsAsBuilt = asBuilt.stream().map(asBuiltAsset -> ContractAgreement
+                        .builder()
+                        .type(ContractType.ASSET_AS_BUILT)
+                        .contractAgreementId(asBuiltAsset.getContractAgreementId())
+                        .id(asBuiltAsset.getId())
+                        .created(Instant.now())
+                        .build()).toList();
+
+                List<ContractAgreement> contractAgreementIdsAsPlanned = asPlanned.stream().map(asPlannedAsset -> ContractAgreement
+                        .builder()
+                        .type(ContractType.ASSET_AS_BUILT)
+                        .contractAgreementId(asPlannedAsset.getContractAgreementId())
+                        .id(asPlannedAsset.getId())
+                        .created(Instant.now())
+                        .build()).toList();
+
+                List<ContractAgreement> mergedAgreements = Stream.concat(contractAgreementIdsAsBuilt.stream(), contractAgreementIdsAsPlanned.stream())
+                        .toList();
+
+                contractService.saveAll(mergedAgreements);
+
+
+            } catch (Exception exception) {
+                log.error("Failed to insert contracts: ", exception);
+            }
+            log.info("on ApplicationReadyEvent insert into contracts successfully done.");
+        });
+        executor.shutdown();
+    }
+
 
 }