+++ tx-backend/src/main/resources/db/migration/R__create_contract_agreement_view.sql
@@ -1,11 +1,10 @@
-create or replace view contract_agreement_view as
-SELECT *
-FROM ((SELECT assets_as_built.id, contract_agreement_id, 'assets_as_built' as asset_type, created
-       FROM assets_as_built
-       where contract_agreement_id is not null)
-      UNION ALL
-      (SELECT assets_as_planned.id, contract_agreement_id, 'assets_as_planned' as asset_type, created
-       FROM assets_as_planned
-       where contract_agreement_id is not null)) results
-ORDER BY created DESC;
+-- Drop the view if it exists
+DROP VIEW IF EXISTS contract_agreement_view;
 
+-- Create the table
+CREATE TABLE contract_agreement (
+                                    id varchar(255) PRIMARY KEY,
+                                    contract_agreement_id varchar(255),
+                                    type VARCHAR(255),
+                                    created TIMESTAMP
+);