+++ CHANGELOG.md
@@ -15,12 +15,14 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - #753 Refactored message history in notification detail view
 - XXX updated local deployment documentation
 - #1037 extended autocomplete api by contractAgreementId
+- #985 Added function to save Contracts based on notification contractAgreementIds into the database
 
 ### Added
 - #832 added policymanagement list view, creator and editor
 - #737 Added concept: Contract table -> parts link action
 - XXX Added interceptor to EdcRestTemplates to log requests
 - #915 Added section to documentation: EDC-BPN configuration
+- #1037 Added link from contracts view to the corresponding filtered part table view
 
 ### Removed
 
@@ -30,6 +32,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 ### Changed
 
 - XXX updated JsonSchemaTest now the test pulls the latest version of the json file
+- XXX deactivated a test class in tx-backend which behaved undesirably
 
 ### Added